Closes #2: Cleanup errors maksing other errors
When we fail to create a resource, the cleanup is triggered, and any
error happening on the cleanup will mask the original error from the
resource creation.

This patch fixes this by preventing any error during the cleanup from
stopping the cleanup operation.  We'll just be logging the errors
instead.